+++ app/lab/murajaah-at-taisir/lib/fetchQuranPage.ts
@@ -11,7 +11,7 @@ type Data = {
   edition: Edition;
 };
 
-type Ayah = {
+export type Ayah = {
   number: number;
   text: string;
   surah: Surah;
@@ -33,8 +33,8 @@ type Surah = {
   numberOfAyahs: number;
 };
 
-type Surahs = {
-  [key: string]: SurahN;
+export type Surahs = {
+  [key: number]: SurahN;
 };
 
 type SurahN = {