+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.10.4] - 2025-11-8
+
+### Fixed
+
+- Improved ayah display logic and layout shifting in murajaah-at-taisir, with refactored code for better readability. See details in [app/lab/murajaah-at-taisir/CHANGELOG.md](app/lab/murajaah-at-taisir/CHANGELOG.md).
+
 ## [v0.10.3] - 2025-11-5
 
 ### Added